# api-gateway/docs/swagger.yaml
definitions:
  model.Cast:
    properties:
      age:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  model.LoginResponse:
    description: LoginResponse information with token,message
    properties:
      message:
        example: login success
        type: string
      token:
        example: asdg123rgfsd3fs51
        type: string
    type: object
  model.Movie:
    description: Movie information with id, name, description, screen_date and cast
    properties:
      cast:
        items:
          $ref: '#/definitions/model.Cast'
        type: array
      description:
        example: movie description
        type: string
      id:
        example: 1
        type: integer
      name:
        example: movie name
        type: string
      screen_date:
        example: 01-01-1999
        type: string
    type: object
  model.SignupResponse:
    description: SignupResponse information with success,username,message
    properties:
      message:
        example: signup success
        type: string
      success:
        example: true
        type: boolean
      username:
        example: username
        type: string
    type: object
  model.User:
    description: User information with username,password
    properties:
      password:
        example: Hello
        type: string
      username:
        example: Hello1
        type: string
    type: object
  util.JsonResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
info:
  contact:
    email: nilenon@gmail.com
    name: Matas Paosriwong
  description: This is a my demo microservice for experimental.
  title: Demo-microservice
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User date to be login
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.JsonResponse'
      summary: Login
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Signup user
      parameters:
      - description: User date to be signup
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/model.SignupResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SignupResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.JsonResponse'
      summary: Signup
      tags:
      - auth
  /movies/:
    get:
      consumes:
      - application/json
      description: Get All Movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Movie'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.JsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.JsonResponse'
      summary: GetMovies
      tags:
      - movies
  /movies/:id:
    get:
      consumes:
      - application/json
      description: Get Movie By ID
      parameters:
      - description: id of movie to be get
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Movie'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.JsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.JsonResponse'
      summary: GetMovieByID
      tags:
      - movies
swagger: "2.0"